+++ packages/react-dom/npm/server.browser.js
@@ -1,0 +1,7 @@
+'use strict';
+
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./cjs/react-dom-server.browser.production.min.js');
+} else {
+  module.exports = require('./cjs/react-dom-server.browser.development.js');
+}
