+++ packages/react-dom/npm/server.js
@@ -1,0 +1,7 @@
+'use strict';
+
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./cjs/react-dom-server.node.production.min.js');
+} else {
+  module.exports = require('./cjs/react-dom-server.node.development.js');
+}
