+++ packages/react-dom/npm/test-utils.js
@@ -1,0 +1,7 @@
+'use strict';
+
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./cjs/react-dom-test-utils.production.min.js');
+} else {
+  module.exports = require('./cjs/react-dom-test-utils.development.js');
+}
