+++ packages/react-dom/server.browser.js
@@ -1,7 +1,4 @@
 'use strict';
 
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-dom-server.browser.production.min.js');
-} else {
-  module.exports = require('./cjs/react-dom-server.browser.development.js');
-}
+// TODO: Move the source into packages
+module.exports = require('ReactDOMServerBrowserEntry');
