+++ packages/react-dom/test-utils.js
@@ -1,7 +1,4 @@
 'use strict';
 
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-dom-test-utils.production.min.js');
-} else {
-  module.exports = require('./cjs/react-dom-test-utils.development.js');
-}
+// TODO: Move the source into packages
+module.exports = require('ReactTestUtilsEntry');
