+++ packages/react-dom/unstable-native-dependencies.js
@@ -1,7 +1,4 @@
 'use strict';
 
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-dom-unstable-native-dependencies.production.min.js');
-} else {
-  module.exports = require('./cjs/react-dom-unstable-native-dependencies.development.js');
-}
+// TODO: Move the source into packages
+module.exports = require('ReactDOMUnstableNativeDependenciesEntry');
