+++ packages/react-noop-renderer/npm/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('./cjs/react-noop-renderer.development.js');
