+++ packages/react-reconciler/index.js
@@ -1,7 +1,4 @@
 'use strict';
 
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-reconciler.production.min.js');
-} else {
-  module.exports = require('./cjs/react-reconciler.development.js');
-}
+// TODO: Move the source into packages
+module.exports = require('ReactFiberReconciler');
