+++ packages/react-test-renderer/index.js
@@ -1,7 +1,4 @@
 'use strict';
 
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-test-renderer.production.min.js');
-} else {
-  module.exports = require('./cjs/react-test-renderer.development.js');
-}
+// TODO: Move the source into packages
+module.exports = require('ReactTestRendererFiberEntry');
