+++ packages/react-test-renderer/shallow.js
@@ -1,7 +1,4 @@
 'use strict';
 
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-test-renderer-shallow.production.min.js');
-} else {
-  module.exports = require('./cjs/react-test-renderer-shallow.development.js');
-}
+// TODO: Move the source into packages
+module.exports = require('ReactShallowRendererEntry');
